app/requirements.txt
Flask==1.0.2
numpy==1.16.3
opencv-python==4.0.1.23
Pillow==8.2.0
redis==3.2.1
scipy==1.6.3
PyYAML==5.4.1
